refactor(schema): extract qaPairRespondent preview to preview/ helper
Mirrors the article/organigramNode/responsibility preview pattern — moves the
select + prepare (incl. answerSnippet + __all__ sentinel) out of qaBlock.ts into
src/preview/qa-pair-respondent-preview.ts. Behavior-preserving.
